requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coverage==5.3
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==2.0.0
kombu==4.6.11
pytz==2020.1
redis==3.5.3
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
zipp==3.2.0
